Add collection options
Currently heroics doesn't support passing additional parameters for
links that return collections. Having the ability to pass collection options are
useful a number of reasons. Mainly, for my use-case I want to use it for
pagination, however, I can see other JSON API's using it for filtering
collections.

This change works by inspecting the `rel` attribute on the link schema
to determine if the link returns a collection.  If `"rel"` contains the value
`instances` this change will add an additional parameter to the function
signature.

Here is an example of the output from the Heroics client generator using
this pull request:

    # Show all sample resources
    #
    # @param collection_options: additional collection options to pass with the request
    def list(collection_options = {})
      @client.another_resource.list(collection_options)
    end

I think this is a useful feature, hopefully it can make it upstream :)
Please let me know what you think.